Add append_path
A function that appends new nodes to a tree of string slices. With it,
parsing stdin should be straightforward: just call append_path for
every line.

I've gone through multiple broken revisions and fixed many interesting
compilation errors in order to arrive at a working version, which
highlights an opportunity for a workflow change, coming soon.