Rewrite print_tree to use an iterator
Adds a `Tree.lines()` method that returns an iterator returning
strings that make up a formatted representation of the tree. With
this, the implementation of `print_tree` becomes trivial. But it's
now also more reusable, as we'll now be able to use the iterator to
render a Cursive view of the tree.

This implementation is based on the "bad" (TBH not so bad...) approach
from Amos Wenger's article "Recursive Iterators in Rust":
https://fasterthanli.me/articles/recursive-iterators-rust