Add a State struct
In addition to the tree, we need to remember which nodes are collapsed
and which aren't. A node is identified by its coordinates, i.e.,
as a Vec<i32>, and we hold a HashSet of such coordinates.

The `collapsed` field is not yet used, but we can already initialize
it – to some non-empty value to exercise the lines() method (which
we will adapt to support collapsing in the next commit).