React to left/right keys by (un)collapsing specific subtree
We put the State instance inside Cursive's user_data field,
so that events callback can get a mutable reference to it
from there. We also give our SelectView a name and wrap it
in OnEventView so that it can react to keypress events.

The node being collapsed is always the same subtree, regardless
of which node is highlighted in the SelectView. Also, redrawing
the SelectView loses track of the current position. Both things
will be fixed later on.