(Un)collapse currently highlighted node
We add a method Tree::coords to convert a usize (currently
selected in the SelectView) to the node coordinates (Vec<i32>)
as per `collapsed`, and use it in the collapse/uncollapse
callback. Also, rewrite redraw() and both callbacks to use
call_on_name("select") to avoid double borrowing of the
SelectView.